src/mlflow/requirements.txt
mlflow==mlflow-2.17.1
sklearn==1.5.2
black==24.10.0
pyspark==3.5.4
delta-spark==3.2.1
dbt-spark==1.9.0
dbt-core==1.9.1
